garotm - XCode project configuration - #55
