+++ ios/RunOn/Build RunOn_2025-01-01T21-07-02.txt
@@ -1,0 +1,28 @@
+
+Showing All Issues
+
+Build target RunOn of project RunOn with configuration Debug
+
+Ld /Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/RunOn.app/RunOn.debug.dylib normal (in target 'RunOn' from project 'RunOn')
+    cd /Users/garotconklin/garotm/fleXRPL/RunOn/ios/RunOn
+    /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/clang -Xlinker -reproducible -target arm64-apple-ios17.6-simulator -dynamiclib -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/iPhoneSimulator.platform/Developer/SDKs/iPhoneSimulator18.2.sdk -O0 -L/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Intermediates.noindex/EagerLinkingTBDs/Debug-iphonesimulator -L/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Intermediates.noindex/EagerLinkingTBDs/Debug-iphonesimulator -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/PackageFrameworks -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/PackageFrameworks -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/PackageFrameworks -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/PackageFrameworks -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/PackageFrameworks -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/PackageFrameworks -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/PackageFrameworks -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/PackageFrameworks -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/PackageFrameworks -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/PackageFrameworks -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/PackageFrameworks -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/PackageFrameworks -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/PackageFrameworks -F/Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator -filelist /Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Intermediates.noindex/RunOn.build/Debug-iphonesimulator/RunOn.build/Objects-normal/arm64/RunOn.LinkFileList -install_name @rpath/RunOn.debug.dylib -Xlinker -rpath -Xlinker /Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/PackageFrameworks -Xlinker -rpath -Xlinker @executable_path/Frameworks -dead_strip -Xlinker -object_path_lto -Xlinker /Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Intermediates.noindex/RunOn.build/Debug-iphonesimulator/RunOn.build/Objects-normal/arm64/RunOn_lto.o -Xlinker -export_dynamic -Xlinker -no_deduplicate -Xlinker -objc_abi_version -Xlinker 2 -Xlinker -debug_variant -fobjc-link-runtime -fprofile-instr-generate -L/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/lib/swift/iphonesimulator -L/usr/lib/swift -Wl,-no_warn_duplicate_libraries -Wl,-no_warn_duplicate_libraries -Wl,-no_warn_duplicate_libraries -Wl,-no_warn_duplicate_libraries -Wl,-no_warn_duplicate_libraries -Wl,-no_warn_duplicate_libraries -Wl,-no_warn_duplicate_libraries -Wl,-no_warn_duplicate_libraries -Wl,-no_warn_duplicate_libraries -Wl,-no_warn_duplicate_libraries -Wl,-no_warn_duplicate_libraries -Wl,-no_warn_duplicate_libraries -Wl,-no_warn_duplicate_libraries -framework CoreGraphics -framework CoreText -framework Foundation -framework LocalAuthentication -framework Security -framework UIKit -framework Security -framework DeviceCheck -framework CFNetwork -Xlinker -alias -Xlinker _main -Xlinker ___debug_main_executable_dylib_entry_point -Xlinker -no_adhoc_codesign -Xlinker -dependency_info -Xlinker /Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Intermediates.noindex/RunOn.build/Debug-iphonesimulator/RunOn.build/Objects-normal/arm64/RunOn_dependency_info.dat -o /Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Products/Debug-iphonesimulator/RunOn.app/RunOn.debug.dylib -Xlinker -add_ast_path -Xlinker /Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Intermediates.noindex/Alamofire.build/Debug-iphonesimulator/Alamofire.build/Objects-normal/arm64/Alamofire.swiftmodule -Xlinker -add_ast_path -Xlinker /Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Intermediates.noindex/GTMAppAuth.build/Debug-iphonesimulator/GTMAppAuth.build/Objects-normal/arm64/GTMAppAuth.swiftmodule -Xlinker -add_ast_path -Xlinker /Users/garotconklin/Library/Developer/Xcode/DerivedData/RunOn-gjaofjwfyafiwmbmbkgnomilqnnw/Build/Intermediates.noindex/GoogleSignIn.build/Debug-iphonesimulator/GoogleSignInSwift.build/Objects-normal/arm64/GoogleSignInSwift.swiftmodule
+
+ld: warning: Could not find or use auto-linked framework 'CoreAudioTypes': framework 'CoreAudioTypes' not found
+ld: warning: Could not parse or use implicit file '/Applications/Xcode.app/Contents/Developer/Platforms/iPhoneSimulator.platform/Developer/SDKs/iPhoneSimulator.sdk/System/Library/Frameworks/SwiftUICore.framework/SwiftUICore.tbd': cannot link directly with 'SwiftUICore' because product being built is not an allowed client of it
+Undefined symbols for architecture arm64:
+  "_main", referenced from:
+      ___debug_main_executable_dylib_entry_point in command-line-aliases-file
+ld: symbol(s) not found for architecture arm64
+clang: error: linker command failed with exit code 1 (use -v to see invocation)
+
+Could not find or use auto-linked framework 'CoreAudioTypes': framework 'CoreAudioTypes' not found
+
+Could not parse or use implicit file '/Applications/Xcode.app/Contents/Developer/Platforms/iPhoneSimulator.platform/Developer/SDKs/iPhoneSimulator.sdk/System/Library/Frameworks/SwiftUICore.framework/SwiftUICore.tbd': cannot link directly with 'SwiftUICore' because product being built is not an allowed client of it
+
+Undefined symbol: _main
+
+Linker command failed with exit code 1 (use -v to see invocation)
+
+
+
+Build failed    1/1/25, 9:07 PM    7.3 seconds
