--- ios/RunOn/RunOn/ContentView.swift
@@ -1,24 +1,0 @@
-//
-//  ContentView.swift
-//  RunOn
-//
-//  Created by Garot Conklin on 12/31/24.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-#Preview {
-    ContentView()
-}
