+++ ios/RunOn/RunOn/Features/Authentication/AuthenticationView.swift
@@ -1,0 +1,45 @@
+import SwiftUI
+import GoogleSignInSwift
+
+struct AuthenticationView: View {
+    @EnvironmentObject var viewModel: AuthenticationViewModel
+    
+    var body: some View {
+        VStack(spacing: 24) {
+            Spacer()
+            
+            VStack(spacing: 12) {
+                Text("Welcome to RunOn")
+                    .font(.system(size: 32, weight: .bold))
+                    .multilineTextAlignment(.center)
+                
+                Text("Sign in to get started")
+                    .font(.system(size: 16))
+                    .foregroundColor(.secondary)
+            }
+            
+            Spacer()
+            
+            GoogleSignInButton(scheme: .dark, style: .wide, state: .normal) {
+                viewModel.signInWithGoogle()
+            }
+            .frame(height: 48)
+            .padding(.horizontal, 24)
+            
+            Spacer()
+                .frame(height: 48)
+        }
+        .padding()
+        .alert("Sign In Error",
+               isPresented: Binding(
+                get: { viewModel.error != nil },
+                set: { if !$0 { viewModel.error = nil } }
+               )) {
+            Button("OK") {
+                viewModel.error = nil
+            }
+        } message: {
+            Text(viewModel.error ?? "An unknown error occurred")
+        }
+    }
+}
