--- ios/RunOn/RunOn/RunOnApp.swift
@@ -1,17 +1,0 @@
-//
-//  RunOnApp.swift
-//  RunOn
-//
-//  Created by Garot Conklin on 12/31/24.
-//
-
-import SwiftUI
-
-@main
-struct RunOnApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
