+++ ios/tmp.xcconfig
@@ -1,0 +1,1 @@
+GOOGLE_CLIENT_ID_ENV = 582143338826-tg2rsh26otaokqor4lgpo9fik3q5oo6b.apps.googleusercontent.com
